자잘한 리팩토링
