+++ src/App.js
@@ -7,5 +7,4 @@ function App() {
     </div>
   );
 }
-
 export default App;
