+++ src/pages/SignUpPage.js
@@ -1,39 +1,26 @@
-import React, { useEffect, useState } from 'react';
-import { useCookies } from 'react-cookie';
+import React, { useState } from 'react';
 import axios from 'axios';
 import qs from 'querystring';
 import styled from 'styled-components';
-const loginURL = 'http://ec2-3-34-5-220.ap-northeast-2.compute.amazonaws.com:8080/auth/login'
-function LoginPage(){
+const signupURL = 'http://ec2-3-34-5-220.ap-northeast-2.compute.amazonaws.com:8080/auth/signup'
+function SignUpPage({setIsSignUpFinished}){
+    const [name, setName] = useState("");
     const [email, setEmail] = useState("");
     const [password, setPassword] = useState("");
-    const [isEmailRemember, setIsEmailRemember] = useState(false);
-    const [cookies, setCookie, removeCookie] = useCookies(['rememberEmail']);
-    useEffect(() => {
-        if(cookies.rememberEmail !== 'undefined'){
-            setEmail(cookies.rememberEmail);
-            setIsEmailRemember(true);
-        }
-    }, []);
+    const handleNameChange = (e) => {
+        setName(e.target.value);
+    }
     const handleEmailChange = (e) => {
         setEmail(e.target.value);
     }
     const handlePasswordChange = (e) => {
         setPassword(e.target.value);
     }
-    const handleEmailRememberButtonClick = (e) => {
-        setIsEmailRemember(e.target.check);
-        if(e.target.check){
-            setCookie('rememberEmail', email, {maxAge: 2000});
-        }
-        else{
-            removeCookie('rememberEmail');
-        }
-    }
-    const handleLoginButtonClick = () => {
+    const handleSignUpButtonClick = () => {
         axios({
-            method: 'post', url: loginURL,
+            method: 'post', url: signupURL,
             data: qs.stringify({
+                name: name,
                 email: email,
                 password: password
             }),
@@ -41,16 +28,22 @@ function LoginPage(){
                 'content-type': 'application/x-www-form-urlencoded'
             }
         })
-        .then ((response) => {
-            setCookie('token', JSON.stringify(response.data));
+        .then (() => {
+            alert('회원가입 성공');
+            setIsSignUpFinished('true');
+            
         })
-        .catch (() => {
-            alert('로그인 실패');
-        })
+        .catch((err) => {
+            alert('회원가입 실패');
+        });
     }
     return(
         <Wrapper>
-            <Title>로그인</Title>
+            <Title>회원가입</Title>
+            <EachField>
+                이름
+                <InputField value={name} onChange={handleNameChange} placeholder="이름을 입력하세요"/>
+            </EachField>
             <EachField>
                 이메일
                 <InputField value={email} onChange={handleEmailChange} placeholder="이메일을 입력하세요"/>
@@ -59,15 +52,10 @@ function LoginPage(){
                 비밀번호
                 <InputField type="password" value={password} onChange={handlePasswordChange} placeholder="비밀번호를 입력하세요"/>
             </EachField>
-            <IsEmailRememberCheckBox>
-                <input type="checkbox" onChange={handleEmailRememberButtonClick} checked={isEmailRemember}/>
-                이메일 저장하기
-            </IsEmailRememberCheckBox>
-            <LoginButton onClick={handleLoginButtonClick}>로그인</LoginButton>
+            <SignUpButton onClick={handleSignUpButtonClick}>회원가입</SignUpButton>
         </Wrapper>
     )
 }
-export default LoginPage;
 const Wrapper = styled.div`
     @import url('https://fonts.googleapis.com/css2?family=Do+Hyeon&display=swap');
     font-family: 'Do Hyeon', sans-serif;
@@ -100,11 +88,7 @@ const InputField = styled.input`
         outline: none;
     }
 `;
-const IsEmailRememberCheckBox = styled.div`
-    position: absolute; 
-    right: 0;
-`;
-const LoginButton = styled.button`
+const SignUpButton = styled.button`
     font-family: 'Do Hyeon', sans-serif;
     font-size: 17px;
     position: absolute;
@@ -122,3 +106,4 @@ const LoginButton = styled.button`
         outline: none;
     }
 `;
+export default SignUpPage;
